[IMP] product_images: details the step to execute the action in Odoo
The first commit was focused on the configuration and was missing the most important part for the users, the step-by-step guide in Odoo.
This commit is only the beginning of this tutorial and still needs a few improvements. That's why you may reach some TODO tags in the document.

task-2558054